requirements.txt
django
django-bootstrap5
djangorestframework
markdown
django-filter
django-rest-swagger
requests